+++ src/com/troblecodings/guilib/ecs/entitys/render/UITexture.java
@@ -17,7 +17,7 @@ import net.minecraftforge.api.distmarker.OnlyIn;
 @OnlyIn(Dist.CLIENT)
 public class UITexture extends UIComponent {
 
-    private final float u, v, mu, mv;
+    private final double u, v, mu, mv;
     private final AbstractTexture texture;
 
     @SuppressWarnings("deprecation")
@@ -29,8 +29,8 @@ public class UITexture extends UIComponent {
         this(texture, 0, 0, 1, 1);
     }
 
-    public UITexture(final ResourceLocation texture, final float u, final float v,
-            final float maxU, final float maxV) {
+    public UITexture(final ResourceLocation texture, final double u, final double v,
+            final double maxU, final double maxV) {
         Minecraft mc = Minecraft.getInstance();
         this.texture = mc.getTextureManager().getTexture(texture);
         this.u = u;
@@ -48,10 +48,10 @@ public class UITexture extends UIComponent {
         final Tesselator tessellator = Tesselator.getInstance();
         final BufferBuilder bufferbuilder = tessellator.getBuilder();
         bufferbuilder.begin(Mode.QUADS, DefaultVertexFormat.POSITION_TEX);
-        bufferbuilder.vertex(0, h, 0).uv(u, mv).endVertex();
-        bufferbuilder.vertex(w, h, 0).uv(mu, mv).endVertex();
-        bufferbuilder.vertex(w, 0, 0).uv(mu, v).endVertex();
-        bufferbuilder.vertex(0, 0, 0).uv(u, v).endVertex();
+        bufferbuilder.vertex(0, h, 0).uv((float)u, (float)mv).endVertex();
+        bufferbuilder.vertex(w, h, 0).uv((float)mu, (float)mv).endVertex();
+        bufferbuilder.vertex(w, 0, 0).uv((float)mu, (float)v).endVertex();
+        bufferbuilder.vertex(0, 0, 0).uv((float)u, (float)v).endVertex();
         tessellator.end();
     }
